Major bug fix for interval != 60 and for command line use.
I had erroneously added a variable window and detrend length based on
the interval argument to calc_e. The change worked for interval=60 but
breaks horribly for interval=1. I have removed the change. Note that the
time window extends for an hour on either side of the input time series
when interval=60 which may be too large for single day processing (i.e.,
even further justification to process 3 consecutive days of data when
processing 1-minute interval data).

Command line usage was broken due to a name change that was not
propagated into the __main__ function (and not exercised until
recently).

Added units to documentation.