Complete retooling of all line-of-sight integration code.
Functionality / compatibility has been maintained (mostly, the only user
significant facing change is that inputs are now PyPositions whereas
before they were xyz ECEF [m]). The changes were made to generalize the
concept of slant integration and enable the use of this common machinery
for other uses.

We've also switched from fixed_quad to quad. Prior, fixed_quad seemed to
provide a more robust approach, but this was a mirage --- with
fixed_quad, we were getting lucky and avoiding NaN regions that were
resulting in the interpolated grids. Some tuning may be required to get
a good accuracy / performance trade-off when using quad.