Checking in (unfinished) GNSS data processing system.
I made a push to improve / enhance the GNSS data processing system
some time back. I have not yet bothered to check in the changes. This
commit includes these changes, but note that a lot of code is now
unclean and a push would have to be made to strengthen the code prior
to actual use.

This commit thus comes with a caveat: use at your own risk for
now. Further testing / validation / cleanup is required.